Added hyperscan fuzzy search
