+++ dbms/tests/queries/0_stateless/00929_multi_match_edit_distance.reference
@@ -1,0 +1,32 @@
+1
+1
+1
+1
+1
+1
+1
+1
+1
+1
+1
+1
+1
+1
+1
+1
+1
+1
+1
+1
+1
+1
+1
+1
+1
+1
+1
+1
+1
+1
+1
+1
